Move tests to corresponding files.
