+++ ui/appui-layout-react/src/test/state/internal/PanelStateHelpers.test.ts
@@ -3,7 +3,12 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import { expect } from "chai";
-import { getPanelMaxSize } from "../../../appui-layout-react/state/internal/PanelStateHelpers";
+import {
+  getPanelMaxSize,
+  insertPanelWidget,
+} from "../../../appui-layout-react/state/internal/PanelStateHelpers";
+import { createNineZoneState } from "../../../appui-layout-react";
+import { addTabs, handleMetaData } from "../../Utils";
 
 describe("getPanelMaxSize", () => {
   it("should use percentage for vertical panel", () => {
@@ -22,5 +27,17 @@ describe("getPanelMaxSize", () => {
       { percentage: 80 }
     );
     expect(sut).to.eq(800);
+  });
+});
+
+describe("insertPanelWidget", () => {
+  it("should throw if `maxWidgetCount` is exceeded", () => {
+    let state = createNineZoneState();
+    state = addTabs(state, ["t1", "t2", "t3"]);
+    state = insertPanelWidget(state, "left", "w1", ["t1"], 0);
+    state = insertPanelWidget(state, "left", "w2", ["t2"], 1);
+    handleMetaData(() =>
+      insertPanelWidget(state, "left", "w3", ["t3"], 2)
+    ).should.throw();
   });
 });
