+++ ui/appui-layout-react/src/test/state/internal/WidgetStateHelpers.test.ts
@@ -6,6 +6,7 @@ import { expect } from "chai";
 import { createNineZoneState } from "../../../appui-layout-react";
 import {
   addFloatingWidget,
+  addPopoutWidget,
   addWidgetState,
   createWidgetState,
   getNewFloatingWidgetBounds,
@@ -60,6 +61,32 @@ describe("addWidgetState", () => {
     handleMetaData(() => addWidgetState(state, "w2", ["t1"])).should.throw(
       "Tab is already in a widget"
     );
+  });
+});
+
+describe("addFloatingWidget", () => {
+  it("should throw if floating widget is already added", () => {
+    let state = createNineZoneState();
+    state = addTabs(state, ["t1", "t2"]);
+    state = addFloatingWidget(state, "fw1", ["t1"]);
+    (() => addFloatingWidget(state, "fw1", ["t2"])).should.throw();
+  });
+
+  it("should set `resizable`", () => {
+    let state = createNineZoneState();
+    state = addTab(state, "t1", { isFloatingWidgetResizable: true });
+    state = addFloatingWidget(state, "fw1", ["t1"]);
+    expect(state.floatingWidgets.byId.fw1.resizable).to.true;
+  });
+});
+
+describe("addPopoutWidget", () => {
+  it("should throw with multiple tabs", () => {
+    let state = createNineZoneState();
+    state = addTabs(state, ["t1", "t2"]);
+    handleMetaData(() =>
+      addPopoutWidget(state, "fw1", ["t1", "t2"])
+    ).should.throw();
   });
 });
